FEAT:"사용자 채팅방 내부 점수 갱신 시스템 Spring Batch"

+++ src/main/java/com/umc/yeongkkeul/domain/mapping/ChatRoomMembership.java
@@ -53,5 +53,5 @@ public class ChatRoomMembership extends BaseEntity {
     private LocalDateTime bannedAt;
 
     @Column
-    private Long userScore;
+    private Double userScore;
 }
